Buffs explorer suit.
I'd like to make its similar (And a bit better.) to armor vest because of it's (suit's) rareness and... well, damn, surely it'll be made of something more durable than your common steel plate carrier. *And it got nerfed to shit on BD a year ago and I'm still mad about that so hey.* Also, makes it able to store weapons in suit slot.